+++ Test-MAME-YM2608/types.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include <stdint.h>
-
-/// Define types
-typedef int32_t		sample;